Refactor import paths and simplify `renderItem` functions in `<ToolbarFilter />`
